View width bug
Width of the view was based on the SNP range returned and this was used
to drive the annotation width.  However, if the first model for a given
gene was narrow, it restricted the display range.  This is fixed and we
uncoupled the annotation range to a wider and recorded width to limit
the number of annotation fetches.